Updates for issue #549
Refactoring concurrent_011 into base rule multiline_simple_structure.
