+++ vsg/rules/__init__.py
@@ -66,6 +66,7 @@ from .token_case_n_token_after_tokens_between_tokens import token_case_n_token_a
 from .existence_of_tokens_which_should_not_occur import existence_of_tokens_which_should_not_occur
 from .multiline_alignment_between_tokens import multiline_alignment_between_tokens
 from .multiline_structure import multiline_structure
+from .multiline_simple_structure import multiline_simple_structure
 from .multiline_constraint_structure import multiline_constraint_structure
 from .multiline_procedure_call_structure import multiline_procedure_call_structure
 from .remove_carriage_return_after_token import remove_carriage_return_after_token
